requirements.txt
albumentations==1.3.1
matplotlib==3.8.1
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.0
Pillow==10.1.0
Pillow==10.2.0
scikit_learn==1.3.2
scipy==1.12.0
seaborn==0.13.2
torch==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
